added apps category, product, storage. Refactoring.
