+++ project_1345/cart/models.py
@@ -1,41 +1,1 @@
-from django.db import models
-from django.db.models import JSONField
-
-
 # Create your models here.
-class Category(models.Model):
-    name = models.CharField(max_length=255)
-    parent = models.ForeignKey("self", on_delete=models.CASCADE, null=True, blank=True)
-    active = models.BooleanField(default=True)
-
-    def __str__(self):
-        return self.name
-
-
-class Product(models.Model):
-    name = models.CharField(max_length=255)
-    price = models.DecimalField(max_digits=10, decimal_places=2)
-    category = models.ForeignKey(Category, on_delete=models.CASCADE)
-    property = models.JSONField()
-    date_added = models.DateField(auto_now_add=True)
-    active = models.BooleanField(default=True)
-
-    def __str__(self):
-        return self.name
-
-
-class Storage(models.Model):
-    product = models.ForeignKey(Product, on_delete=models.CASCADE)
-    quantity = models.IntegerField(default=1)
-    date_added = models.DateField(auto_now_add=True)
-    active = models.BooleanField(default=True)
-
-    def __str__(self):
-        return (
-            f"{self.quantity} {self.product.name}"
-            if self.quantity > 1
-            else self.product.name
-        )
-
-    def __repr__(self):
-        return str(self)
