+++ project_1345/product/models.py
@@ -1,0 +1,16 @@
+from django.db import models
+
+from project_1345.category.models import Category
+
+
+# Create your models here.
+class Product(models.Model):
+    name = models.CharField(max_length=255)
+    price = models.DecimalField(max_digits=10, decimal_places=2)
+    category = models.ForeignKey(Category, on_delete=models.CASCADE)
+    property = models.JSONField()
+    date_added = models.DateField(auto_now_add=True)
+    active = models.BooleanField(default=True)
+
+    def __str__(self):
+        return self.name
