+++ project_1345/project_1345/settings.py
@@ -38,7 +38,9 @@ INSTALLED_APPS = [
     "django.contrib.messages",
     "django.contrib.staticfiles",
     "drf_yasg",
-    "cart.apps.project_1345",
+    "cart",
+    "category",
+    "product",
 ]
 
 MIDDLEWARE = [
