+++ project_1345/storage/admin.py
@@ -1,0 +1,3 @@
+from django.contrib import admin
+
+# Register your models here.
